Vipl.yaml
Viplant is electronic cash and a project to attract investment funds.
This is a combination thanks to which a VIP factory will be built with investors' funds. Vipl token owners will be able to purchase products that such an enterprise will produce, and pay with the Vipl token directly from their crypto wallet. So Vipl will be electronic cash for which you can purchase products, and not just other cryptocurrency.