Create EcoDoge
EcoDoge is a meme coin created in 2025 in response to the global environmental agenda and the growing popularity of sustainable development. Riding the wave of the trend for environmental initiatives, EcoDoge combines elements of humor and seriousness, offering users not only the opportunity to earn money but also to contribute to environmental protection. Every time users conduct transactions, part of the commission is automatically sent to projects to restore forests and clean the oceans. EcoDoge has become a symbol of a new era, where technology and ecology go hand in hand.