Create ATC
Followers can directly participate in a part of the profit of the system by buying ATC tokens, and the system allocates at least 10% of the monthly profit every month to buy ATC tokens and increase its liquidity, and the followers of the system should always try to keep the maximum amount of ATC tokens. to make more profit.