Create ATC.yaml
about jetton : 
The ATC token is the first Iranian token written in FUNC and built on the TON blockchain. AutoTCrypto envisions creating a dual-loyalty token for offering discounts, cashback payments, and allocating a percentage of the system’s profits to its followers.  

Followers can directly share in the system’s profits by purchasing ATC tokens. Each month, the system allocates at least 10% of its monthly profit to buying ATC tokens and increasing its liquidity. Followers are encouraged to hold as many ATC tokens as possible to maximize their earnings.  

ATC is not meant for trading or serving as an intermediary for purchases; rather, it is a token designed to honor AutoTCrypto’s loyal audience and serve as an investment. Loyal followers must exchange their money on reputable exchanges such as Nobitex, Wallex, and others, as currency conversion is not available within the system. Token purchases take place on decentralized exchanges on the blockchain.