Create CHROTHER
CHROTHER – a coin that creates the future. Turning time into value.

CHROTHER is not just a token but a tool for real-world use. It will become part of an ecosystem that includes:

Freelance marketplace – a convenient platform for clients and freelancers where payments are made in CHROTHER.

Quest application – complete tasks, earn coins, and build your capital.

Mining system – an innovative approach that allows mining third-party tokens and converting them into CHROTHER.