Create LIZZARD.yaml
Hello, we want to warn you right away that our coin is owned by 205 people. There are 210 active users on the site.
Also on the site it is written that we do not have any accounts in social networks, because all the names are occupied.
And that's why we have created a website on which we notify our investors.