Create GLUF.yaml
Persian Gluf Token ($GLUF) is a community-driven meme coin built on the TON blockchain, celebrating the power, beauty, and meme energy of the Persian Gulf.

Designed for fun, culture, and decentralization, GLUF blends regional pride with the viral momentum of meme coins to build a powerful community. With unique branding, a locked LP of 500,000,000 GLUF, and exciting future plans, GLUF aims to ride the wave of global crypto attention.

Join the meme storm — from the heart of the Gulf to the decentralized world.

Please make sure you change the original .yaml fields in the accounts/, collections/ or jettons/ directories and leave the auto-generated .json files in the repository root alone. Also please make sure that you do not use ton.api links in your pull request.
Example pull request: