Create ANG3LS.yaml
Collection of NFTs I have and would like to get them verified. This is the Metadata from getgems.io where I created them. I viewed it and entered it in this format because I wasn't sure how to. If you need it to be exactly as the meta data layout just give me some feedback and I can do it again. Thank you for your time and effort