Update meta.json
Aureum (AUR) is more than just a token — it is a vision of stability and trust within the TON ecosystem. Built with a fixed supply of 1 billion tokens and 9 decimals, Aureum combines the classic value of 'gold' with the innovation of blockchain technology. The token is designed to support trading, holding, and community-driven projects, making it a cornerstone for long-term growth and adoption in the TON network.