Please, verify BULL.yaml
We request expedited review due to a serious problem with impersonators who have copied our brand, logo, and images to create fraudulent token copies and now we are tagged as FAKE in TonKeeper wallet. We are a legitimate project with a roadmap for organic growth and expansion within the TON ecosystem.

You can confirm our authenticity and the timeline of impersonation by checking the onchain creation date of our official $BULL token against those using the same ticker.